requirements.txt
appdirs==1.4.3
astroid==2.3.1
attrs==19.3.0
beautifulsoup4==4.8.1
black==19.3b0
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cssselect==1.1.0
Django==2.2.6
djangorestframework==3.10.3
fake-useragent==0.1.11
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
lxml==4.4.1
mccabe==0.6.1
parse==1.12.1
pyee==6.0.0
pylint==2.4.2
pyppeteer==0.0.25
pyquery==1.4.0
pytz==2019.3
requests==2.22.0
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
toml==0.10.0
tqdm==4.36.1
typed-ast==1.4.0
urllib3==1.25.6
w3lib==1.21.0
websockets==8.0.2
wrapt==1.11.2
